requirements.txt
streamlit==1.34.0
llama_cpp_python==0.2.76
